actor/requirements.txt
agavepy
pandas
pandas_bokeh
matplotlib